Day 1, Part 2
The "simple" but perhaps less performant approach
of accounting for spelled out digits doesn't work.
because it turns out that first digits need to be
replaced from the beginning and last digits need to
be replaced from the end. (This wasn't caught by the
test data, just the prod data. Nasty.)

Build recursive firstDigitPart2 and lastDigitPart2
functions that iterate bit by bit through the input,
stopping when they see a digit, or when the start (or
end) of the input spells out a digit. This might be
more performant for long strings because the whole
line usually isn't transformed.