Day 3, Part 1
Extract the part numbers using a regex and a filter that
checks against the signal positions on each line.
Convert the part numbers to longs and calculate the sum.

(Don't return the part numbers as a set, because some
lines in the production input have duplicate part
numbers. Learned that one the hard way.)