Day 3, Part 2
Use the refactored findPartNumbers function to build
an inverse mapping of SymbolPosition to adjacent
PartNumbers (well, the values at least). Filter this
by the symbols that are '*'s, and filter for entries
that have 2 adjacent part numbers. These are the
gears as described in the problem. Calculate the
product of adjacent part numbers for each gear
(the "gear ratio") and then calculate and return the
sum of all gear ratios.