Use Gears.partNumbers to calculate gear product
This requires a couple of steps. Since reducing a set
isn't quite the same as explicitly multiplying each
gear value (different if the two part numbers are the
same!), we need to switch it to a list. This requires
an equalsIgnoringOrder helper method in the test
class.

Once this is done, a quick call to reduce generates
the answer, and the vals in the Gear class can be
removed as well.