Day 5, Part 1
The input file has such large ranges of number that
building an explicit mapping would run out of memory
instantly. So build a class that holds the map
definitions and applies them just in time with a stream
operation and a range test.

Build every mapping in the file this way and apply them
all to each seed mapping using fold.

Don't forget that if the map doesn't have a range that
matches the input it should just return the input.