Day 5, part 2 with good performance
For each range, check if it's continuous or not (it's
continuous if there's an even step from transforming
the beginning of the range to the end of the range). If
it's continuous, only the beginning of the range
needs to be checked. If it's not continuous, split the
range into two (approximately) even halves, and
process each separately. Brute force small ranges to
avoid fiddly split scenarios.

Keep going until there are no more ranges to check, and
return the smallest value encountered.

This approach returns a correct answer in <50ms.