Day 6, Part 1
Build a simple parse method with substringAfter and split,
and combine the two lines with zip.

Calculate the ways to beat the record for the races by
brute force: starting at 1ms button press up to
raceTime-1ms, calculate the distance, and count the
entries that exceed the record distance. Then calculate
the product across all of the races.