Day 7, part 2
Handle wild cards by splitting out the Js, and then
adding the number of Js found to the card with the
highest occurence, and then testing for hand types as
before.

Build a part 2 comparator that compares using part 2
hand strength, and using the part 2 ordinal for cards
in case of a tie-break.

Calculate the overall answer as before.

This approach has the same performance characteristics
as part 1, so it's nice and fast.