Day 8, Part 2
For part 2,the result is way too big (14 trillion+
iterations) to calculate directly.

So, start by using the same algorithm as part 1 with a
different termination condition (ends with Z vs "ZZZ") to
calculate the individual number of steps for each node
that starts with A (there are 6 in the prod data). Then
calculate the cumulative least common multiple of all of
those counts to get the overall answer.

This method only works when the first leg and subsequent
loops for each path have the same length, which they do
for test and production data. Yay. Add assert statements
to ensure this.

This approach calculates an answer in ~40ms.