Day 10, Part 1
Find S and the two pipe segments connected to S, which are the
pipe segments adjacent to X (diagonals don't count) that
connect back to S.

Pick one of these at random and follow it in the one
direction (do this by never going back to the point you've
just visited). Once you reach S you've traced out the entire
loop. The length of the loop / 2 is the answer.