Find the loop segments instead of just the length
When travelling the loop, keep a list of all loop segments
visited so far, not just the last one. We only check against
the last entry in the list, so time complexity is unchanged, and
space complexity is linear.

Return the loop points as a Set instead of a List so that the
order we traverse the loop doesn't affect the result.

Loop length is just the length of the loop.

This refactoring will help make the part 2 solution much easier.