Day 2, Part 2
Use a ray-casting point-in-polygon algorithm to determine if a tile
is inside the main loop. Find the edge of the loop via the
algorithm from part 1.

The ray-casting algorithm is complicated by horizontal segments. A
horizontal segment only counts as an intersection if its beginning
and end go in different vertical directions. There's some fairly
ick if/else logic used to track this.

Once that's handled, consider every tile and count it as inside
the main loop if there's an odd number of intersections with loop
tiles to the left of the current row. Since parts of other loops
can count as inside the main loop, we only need to look at the
tile type if it's part of the main loop, to determine if it's a
qualifying horizontal segment.

This approach calculates the correct answer in ~55ms.