Simplify intersection test
This can be dramatically simplied by just testing for characters that
"cross" the horizontal: the vertical pipe and characters that exit via
the top (or bottom — you can choose arbitrarily, either works).
Segments of the wrong type that are on the path just get ignored.

Doesn't change the run time, but makes the logic simpler.