Day 11, Part 1
Parse the galaxies in the usual way.

Expand the universe by:
- creating TreeSets of x and y values with no rows in them by subtracting
the unique x and y coordinates of galaxies that exist from the set of
possible x and y coordinates. Use TreeSets for a fast lookup for all
values < n.
- Iterating through each point and adding the number of empty columns < X
and the number of empty rows < Y to each coordinate. Use headSet for
performance.

Once the universe is expanded, the answer is simply the manhattan distance
between each point and every other point, summed and divided by 2.