Convert Part 1 to dynamic programming style
Instead of tracking the actual substitutions that
have been made so far, track the parse position in
the string, how far along (if at all) we are in a
run of #s, and how many constraints in the list of
ints have been satisfied so far. This allows much
more overlap when we start caching intermediate
results for part 2.