Clean up logic
Remove the extra '.' character at the end of the
record. This makes the logic at the end of the list
more complicated, because we have to check if a run of
#s is ending at exactly the right time, but the extra
logic isn't that bad if we express it in terms of list
equality compared to example lists.

Remove the extra internal field holding the extra . as
well.