Get ready to solve part 1 with hashing
There's a really easy perfect hash code that will
work for this problem: set the nth bit of the hash
code if the nth character is a #. (No line is more
than 32 characters long, even when transposed. There
are tests that check this with production data.)

This hash code also allows "different by 1
character" testing really cheaply: just xor and
check for a single 1 bit.

Since it's a perfect hash, the hash code being equal
guarantees that the values are equal, so the
part 1 portion of the problem can be solved with
just the hash codes for a pattern.