Day 13, part 2
(Finally) fixed the bug! I was disallowing the previous
solution in the wrong place. The check has to happen inside
findMirrorRowInternal. If it doesn't happen in there, a
potential altered pattern that has two matches (the new one
and the old one), might return the old one. Filtering in
findMirrorRowWithSmudge was not enough. So now I calculate
the old result in findMirrorRowWithSmudge, but pass it into
findMirrorRowInternal to be enforced.

Also: rewrite findMirrorRowInternal to avoid grouping by
hash code. It doesn't make the method any simpler or any
faster (it just ended up doing group comparisons against
the lists instead of an equivalent number of comparisons
against the ungrouped list of hash codes).

This approach calculates the correct answer in ~66ms.