Day 17, Part 1
Combine the special movement rule (moving up to 3
spaces in a straight line before turning) with a
shortest path search by building a derived graph that
captures the movement of an entire sequence of 1-3
moves in a single weighted edge, with point +
direction nodes. Then run Djikstra's algorithm on
that to get the answer.