Eliminate up-front construction of the graph
Now that Djikstra's is extracted into a utility function, it
explicitly requires the neighbour lookup for a node to be
performed via a lambda, so it's now a lot more natural for
this to be performed just in time via a function.

Refactor the buildGraph code into a findNeighbours function
that is called from the Djikstra's loop. Test this function
directly instead of via buildGraph functions.

The resulting code is shorter, easier to understand,
and is 200-300ms faster in calculating the result because
the entire graph doesn't need to be constructed up front.

Win.