Day 18, Part 2
Reimplement completely for large inputs.
Turn the dig instructions into a polygon (i.e., a
list of vertices) instead of a set of border points.

Then use the shoelace theorem to calculate the
polygon's area. This is not quite the number of
points in the polygon + its area because this
polygon has a border with real thickness.

Use Pick's Theorem (https://en.wikipedia.org/wiki/Pick%27s_theorem)
to sort out the difference, solving for the
interior points using the area and border points.

(The number of border points is the length of the
perimeter).

Add the interior and border point counts for the
answer.

Now that part1 is fast enough, to handle part 2, write a function
that turns hex codes into part 1 dig instructions. Use the output
as input to calculate an answer with part 1's method.

This approach finds the correct answers in ~30ms.