Day 19, Part 2
Solve the second part efficiently by finding all the paths
to acceptance ("A") through the workflows, via a depth first
search. Conditional clauses are a bit tricky, because either
the clause will apply if the branch was taken, or the
_inverse_ of the clause will apply if it wasn't.

Convert each path into an IntRange of acceptable values
for each of the 4 xmas fields. Then for each possibility
multiply the size of each range together, and return the
overall sum for all the possibilities.

This could be used to check parts for part 1 — just check the
part scores against all of the ranges in turn, but why bother.
The post-DFS processing step could probably be done in the DFS,
but .... meh.

This approach calculates the correct answer in ~70ms.