Remove nodes parameter to djikstras function
Since we're using a default value in tenativeDistances,
if the start nodes are added to the priority queue at the
beginning, then no other nodes need to be added (they get
added/overridden later when they're found in the
neighbours function anyway). This is the only use of
nodes in the function, so it's not actually needed.

Removing it will make the algorithm much more useful in
infinite grid situations.

Change function callers to match.