Day 21, Part 2 (Finally!)
Modified the original approach to work on an infinte grid
by changing the bounds checking, and modifying the
Djikstra's algorithm helper to have a cost limit.

With a lot of help from the forum, figured out that the
result is a quadratic polynomial for the prod data. Used a
borrowed Lagrange formula to calculate the roots (and
checked on wolfram alpha so I know it's calculating the
right roots). Then plug in steps/garden size into the
formula for the answer.

This approach calculates the right answer in ~500ms.