Day 20, Part 2
This solution only works for the production input, and not
the test. After some hints from the message board, I looked
at the module graph, and determined that there is one module
(xm in my case) that sends to rx, and 4 modules that send to
xm. They are all conjunction modules. When the 4 conjunction
modules all send high at the same time, the last one will send
low to rx.

Count the number of button presses needed to make each of the
4 modules emit a high pulse, and then compute the Least Common
Multiple of that.

This approach finds the correct answer in just under 100ms.