Day 24, Part 1
Compute intersection using intercept formulas
hand derived from y = mx+b. M is easy to
calculate because the deltas for each hailstone are
provided, so you need to remember "rise over run".
Once you have M, plug x into y = mx + b to calculate
b = y - mx.

Once the slope and intercept are calculated for each
line, the intersection coordinates can be derived
because we know that x1 = x2 and y1 = y2 at the
intersection point (because after all, they
intersect there!). So we can put the two mx + b
equations equal to each other, replace x2 with x1,
and rearrange the equation to solve for x1. This
looks like intersectX = (b2 - b1)/(m1 -m2).

Once we have the x coordinate of the intersection
point we find the y coordinate with y = mx + b.

Find if the intersection is in the future for each
point by comparing the sign of (intersectionX - x)
with the sign of deltaX.

Then test to see if the intersection point is in
bounds, et voilà.

Compare each possible combination
of 2 hailstones and count the legal intersections.

(This *sounds* easy, but it was absolute maddening,
because I wasted an entire day on a *lot* of messing
around with determinants and other methods of
calculating intesections that were passing the tests
but giving wrong answers for production data. Grr.