Build findIntersection3D
Create findIntersectionXZ: it's the same as
findIntersectionXY except that it works on the XZ
plane.

Combine the two to build findIntersection3D: if the
lines intersect on the XY and XZ plane then they
intersect in three dimensions (check that the X
coordinates from the XY and XZ matches are equal as
a consistency check).

This version returns whole number values, and also
returns the time that the rays intersect (as that
will be useful solving part 2).