Day 23, Part 2
Optimise the graph by only keeping nodes that are
decision points — nodes where there is a meaningful
choice about where to go. Weight the edges with the
number of tiles that would have been travelled in
the original graph to get the a correct answer.

This is.... not really very fast. It answers part 2
in ~35 seconds. Which is disappointing, but fast
enough to provide an answer. Boo.