Remove legacy readInput function
Production code reads (and automatically
downloads, if necessary) input from the
Puzzle class, which means data files don't
need to be checked in. This is good because
AoC rules ask you not to share data files.

Test files can actually use the same
function, so switch to that usage and
remove the legacy readInput function.