Change naming convention
Giving every solution class the same name in a
different package caused a lot of problems:
IntelliJ's refactoring got confused and rename
refactorings corrupted filenames. Also,
automatically jumping to test classes didn't work
very well.

But having sample code in a skeleton file that
can be pasted as-is still feels like a win.

The solution is to name the class Solution in the
skeleton file, but then rename it to DayXXSolution
after its been pasted to the actual package.

Rename the existing classes and tests to match
this convention.

Also clean up the base classes: XXXSolution
classes that extend Puzzle doesn't make sense.